chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` and `package-lock.json` files. It adds the following dependencies:
- axios@1.7.2
- crypto-js@4.2.0

It also updates the version of `next` to 14.2.4 and `react` to 18. Additionally, it updates the dev dependencies `eslint` to version 8 and `eslint-config-next` to version 14.2.4. Finally, it updates `postcss` to version 8 and `tailwindcss` to version 3.4.1.

Please review and merge this commit.